[2024.04.04] 갤러리 등록 추가

+++ backend/src/main/java/com/backend/entity/File.java
@@ -20,7 +20,7 @@ public class File extends BaseEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int fno;
+	private long fno;
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "board_bno")
